third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-inside-expected.txt

Testing parsing of the shape-inside property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getCSSText("shape-inside", "auto") is "auto"
PASS getComputedStyleValue("shape-inside", "auto") is "auto"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px)"
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 5px)"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px)") is "inset-rectangle(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px)") is "inset-rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px)"
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 5px)"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getCSSText("shape-inside", "circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS getComputedStyleValue("shape-inside", "circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS getCSSText("shape-inside", "ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("shape-inside", "ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS getCSSText("shape-inside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("shape-inside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("shape-inside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("shape-inside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("shape-inside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("shape-inside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("shape-inside", "content-box") is "content-box"
PASS getComputedStyleValue("shape-inside", "content-box") is "content-box"
PASS getCSSText("shape-inside", "padding-box") is "padding-box"
PASS getComputedStyleValue("shape-inside", "padding-box") is "padding-box"
PASS getCSSText("shape-inside", "border-box") is "border-box"
PASS getComputedStyleValue("shape-inside", "border-box") is "border-box"
PASS getCSSText("shape-inside", "margin-box") is "margin-box"
PASS getComputedStyleValue("shape-inside", "margin-box") is "margin-box"
PASS getCSSText("shape-inside", "outside-shape") is "outside-shape"
PASS getComputedStyleValue("shape-inside", "outside-shape") is "outside-shape"
PASS removeBaseURL(getCSSText("shape-inside", "url('image')")) is "url(image)"
PASS removeBaseURL(getComputedStyleValue("shape-inside", "url('image')")) is "url(image)"
PASS getCSSText("shape-inside", "calc()") is ""
PASS getComputedStyleValue("shape-inside", "calc()") is "auto"
PASS getCSSText("shape-inside", "none") is ""
PASS getComputedStyleValue("shape-inside", "none") is "auto"
PASS getCSSText("shape-inside", "rectangle()") is ""
PASS getComputedStyleValue("shape-inside", "rectangle()") is "auto"
PASS getCSSText("shape-inside", "rectangle(10px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px)") is "auto"
PASS getCSSText("shape-inside", "rectangle(10px, 10px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 10px)") is "auto"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px)") is "auto"
PASS getCSSText("shape-inside", "rectangle(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px 20px 30px 40px)") is "auto"
PASS getCSSText("shape-inside", "rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS getComputedStyleValue("shape-inside", "rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "auto"
PASS getCSSText("shape-inside", "inset-rectangle()") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle()") is "auto"
PASS getCSSText("shape-inside", "inset-rectangle(10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px)") is "auto"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 10px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 10px)") is "auto"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px)") is "auto"
PASS getCSSText("shape-inside", "inset-rectangle(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px 20px 30px 40px)") is "auto"
PASS getCSSText("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS getComputedStyleValue("shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "auto"
PASS getCSSText("shape-inside", "circle()") is ""
PASS getComputedStyleValue("shape-inside", "circle()") is "auto"
PASS getCSSText("shape-inside", "circle(10px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px)") is "auto"
PASS getCSSText("shape-inside", "circle(10px, 20px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px, 20px)") is "auto"
PASS getCSSText("shape-inside", "circle(10px 20px 30px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px 20px 30px)") is "auto"
PASS getCSSText("shape-inside", "circle(10px, 20px, 30px, 40px)") is ""
PASS getComputedStyleValue("shape-inside", "circle(10px, 20px, 30px, 40px)") is "auto"
PASS getCSSText("shape-inside", "ellipse()") is ""
PASS getComputedStyleValue("shape-inside", "ellipse()") is "auto"
PASS getCSSText("shape-inside", "ellipse(10px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px)") is "auto"
PASS getCSSText("shape-inside", "ellipse(10px, 20px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px, 20px)") is "auto"
PASS getCSSText("shape-inside", "ellipse(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px, 20px, 30px)") is "auto"
PASS getCSSText("shape-inside", "ellipse(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("shape-inside", "ellipse(10px 20px 30px 40px)") is "auto"
PASS getCSSText("shape-inside", "polygon()") is ""
PASS getComputedStyleValue("shape-inside", "polygon()") is "auto"
PASS getCSSText("shape-inside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is "auto"
PASS getCSSText("shape-inside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is "auto"
PASS getCSSText("shape-inside", "polygon(nonzero)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(nonzero)") is "auto"
PASS getCSSText("shape-inside", "polygon(evenodd)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(evenodd)") is "auto"
PASS getCSSText("shape-inside", "polygon(10px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(10px)") is "auto"
PASS getCSSText("shape-inside", "polygon(nonzero,10px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(nonzero,10px)") is "auto"
PASS getCSSText("shape-inside", "polygon(evenodd,12px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(evenodd,12px)") is "auto"
PASS getCSSText("shape-inside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is ""
PASS getComputedStyleValue("shape-inside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is "auto"
PASS getParentAndChildComputedStylesString("shape-inside", "auto", "rectangle(10px, 20px, 30px, 40px)") is "parent: auto, child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("shape-inside", "outside-shape", "rectangle(10px, 20px, 30px, 40px)") is "parent: outside-shape, child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("shape-inside", "rectangle(10px, 20px, 30px, 40px)", "initial") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: auto"
PASS getParentAndChildComputedStylesString("shape-inside", "rectangle(10px, 20px, 30px, 40px)", "") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: auto"
PASS getParentAndChildComputedStylesString("shape-inside", "rectangle(10px, 20px, 30px, 40px)", "inherit") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("shape-inside", "", "inherit") is "parent: auto, child: auto"
PASS getParentAndChildComputedStylesString("shape-inside", "auto", "inherit") is "parent: auto, child: auto"
PASS successfullyParsed is true

TEST COMPLETE
